pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ncmdump-py"
authors = [
    {name = "ww-rm", email = "ww-rm@qq.com"},
]
description = "Dump ncm files to mp3 or flac files."
requires-python = ">=3.7"
dependencies = [
    "mutagen",
    "Pillow",
    "pycryptodome",
    "rich",
]
dynamic = ["version", "readme"]

[project.urls]
"Homepage" = "https://github.com/ww-rm/ncmdump-py"
"Bug Tracker" = "https://github.com/ww-rm/ncmdump-py/issues"

[tool.setuptools]
packages = ["ncmdump"]

[tool.setuptools.dynamic]
version = {attr = "ncmdump.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}
